fix: correctly handle derivations in `configFiles`
Derivations are not paths, but pathlike so `builtins.isPath` reports
`false` for them. However, what we actually want to check for is
whether the given value is path-like or not. So check if the value is a
string and if not, assume it some path like item (either a path or a
derivation).